Update to Patch 6.4.

+++ src/test/java/teemowork/tool/riot/ChampionDefinition.java
@@ -309,9 +309,9 @@ class ChampionDefinition implements Comparable<ChampionDefinition> {
 
         public List<Float> cooldown;
 
-        public List<Integer> cost;
+        public List<Float> cost;
 
-        public List<Integer> range;
+        public List<Float> range;
 
         public String costType;
